LayoutTests/media/media-vp8-webm-error-offscreen.html

<!DOCTYPE html>
<html>
<head>
    <title>media-vp8-webm-error-offscreen</title>
    <script src=video-test.js></script>
    <script>
        window.addEventListener('load', event => {
            run('video = document.createElement("video")');
            video.muted = 1;

            video.onerror = (event) => {
                consoleWrite("received error event");
                endTest();
            };

            video.oncanplay = () => {
                video.play();
            };

            video.onloadedmetadata = (event) => {
                consoleWrite(`metadataloaded: ${video.videoWidth}x${video.videoHeight}`);
            }

            run('video.src = "content/garbage.webm"');
        });

    </script>
</head>
<body>
</body>
</html>
