LayoutTests/media/media-vp8-webm-error-offscreen-expected.txt

RUN(video = document.createElement("video"))
RUN(video.src = "content/garbage.webm")
metadataloaded: 320x240
received error event
END OF TEST
